requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
tzdata==2025.2
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.3.2
pytest-cov==4.1.0

pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-benchmark==4.0.0
django-debug-toolbar==4.2.0

responses==0.24.1
freezegun==1.3.1

flake8==6.1.0
black==23.12.1
isort==5.13.2
pylint==3.0.3

pillow==10.1.0